Corrected detekt report issues
