+++ canarinho/src/androidTest/kotlin/br/com/concrete/yosef/canarinho/component/CanarinhoComponentTextField.kt
@@ -3,7 +3,9 @@ package br.com.concrete.yosef.canarinho.component
 import android.support.test.espresso.Espresso.onView
 import android.support.test.espresso.action.ViewActions.typeText
 import android.support.test.espresso.assertion.ViewAssertions.matches
-import android.support.test.espresso.matcher.ViewMatchers.*
+import android.support.test.espresso.matcher.ViewMatchers.isAssignableFrom
+import android.support.test.espresso.matcher.ViewMatchers.withText
+import android.support.test.espresso.matcher.ViewMatchers.isDisplayed
 import android.support.test.rule.ActivityTestRule
 import android.support.test.runner.AndroidJUnit4
 import android.widget.EditText
