+++ canarinho/src/main/kotlin/br/com/concrete/yosef/canarinho/api/property/MaskPropertyCommand.kt
