+++ core/src/androidTest/kotlin/br/com/concrete/yosef/component/ElementGroupComponentTest.kt
@@ -1,4 +1,4 @@
-package br.com.concrete.yosef.ui.component
+package br.com.concrete.yosef.component
 
 import android.support.test.espresso.Espresso
 import android.support.test.espresso.assertion.ViewAssertions
