+++ core/src/androidTest/kotlin/br/com/concrete/yosef/property/text/TextCommandTest.kt
@@ -1,4 +1,4 @@
-package br.com.concrete.yosef.ui.property.text
+package br.com.concrete.yosef.property.text
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
@@ -7,8 +7,8 @@ import android.widget.ImageView
 import android.widget.LinearLayout
 import android.widget.TextView
 import br.com.concrete.yosef.afterLayout
-import br.com.concrete.yosef.api.property.text.TextSizeCommand
-import br.com.concrete.yosef.api.property.text.TextSizeCommand.Companion.TEXT_SIZE
+import br.com.concrete.yosef.api.property.text.TextCommand
+import br.com.concrete.yosef.api.property.text.TextCommand.Companion.TEXT
 import br.com.concrete.yosef.entity.DynamicProperty
 import org.junit.Assert.assertEquals
 import org.junit.Before
@@ -18,7 +18,7 @@ import org.junit.rules.ExpectedException
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class TextSizeCommandTest {
+class TextCommandTest {
 
     @Rule
     @JvmField
@@ -26,45 +26,44 @@ class TextSizeCommandTest {
 
     private val context = InstrumentationRegistry.getTargetContext()!!
 
-    private lateinit var textSizeCommand: TextSizeCommand
+    private lateinit var textCommand: TextCommand
 
     private lateinit var parent: ViewGroup
 
     @Before
     fun setUp() {
-
         parent = LinearLayout(context).apply {
             orientation = LinearLayout.VERTICAL
         }
-        textSizeCommand = TextSizeCommand()
-
+        textCommand = TextCommand()
     }
 
     @Test
-    fun renderingTextViewShouldApplyTextSize() {
-        val dynamicProperty = DynamicProperty(TEXT_SIZE, "dimen", "25")
+    fun renderingTextViewShouldApplyText() {
+        val dynamicProperty = DynamicProperty(TEXT, "string", "Teste texto")
 
         val textView = TextView(parent.context)
-        textSizeCommand.apply(textView, dynamicProperty)
+        textCommand.apply(textView, dynamicProperty)
 
         parent.addView(textView)
 
         textView.afterLayout {
-            assertEquals(25f, textView.textSize)
+            assertEquals(textView.text.toString(), dynamicProperty.value)
         }
+
     }
 
     @Test
     fun renderingImageViewShouldThrow() {
-        val dynamicProperty = DynamicProperty(TEXT_SIZE, "dimen", "25")
+        val dynamicProperty = DynamicProperty(TEXT, "string", "Teste texto")
 
         val imageView = ImageView(parent.context)
 
         exceptionRule.expect(IllegalArgumentException::class.java)
-        exceptionRule.expectMessage("The property does not support setting text size" +
-                " for the type ${imageView.javaClass.name}")
+        exceptionRule.expectMessage("The value (${dynamicProperty.value}) for the $TEXT " +
+                "property is not compatible with ${imageView.javaClass.name}")
 
-        textSizeCommand.apply(imageView, dynamicProperty)
+        textCommand.apply(imageView, dynamicProperty)
     }
 
 }
