+++ core/src/androidTest/kotlin/br/com/concrete/yosef/property/text/TextSizeCommandTest.kt
@@ -1,15 +1,16 @@
-package br.com.concrete.yosef.ui.property.text
+package br.com.concrete.yosef.property.text
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
 import android.view.ViewGroup
+import android.widget.ImageView
 import android.widget.LinearLayout
 import android.widget.TextView
 import br.com.concrete.yosef.afterLayout
-import br.com.concrete.yosef.api.property.text.TextStyleCommand
-import br.com.concrete.yosef.api.property.text.TextStyleCommand.Companion.TEXT_STYLE
+import br.com.concrete.yosef.api.property.text.TextSizeCommand
+import br.com.concrete.yosef.api.property.text.TextSizeCommand.Companion.TEXT_SIZE
 import br.com.concrete.yosef.entity.DynamicProperty
-import org.junit.Assert.assertTrue
+import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -17,7 +18,7 @@ import org.junit.rules.ExpectedException
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class TextStyleCommandTest {
+class TextSizeCommandTest {
 
     @Rule
     @JvmField
@@ -25,7 +26,7 @@ class TextStyleCommandTest {
 
     private val context = InstrumentationRegistry.getTargetContext()!!
 
-    private lateinit var textStyleCommand: TextStyleCommand
+    private lateinit var textSizeCommand: TextSizeCommand
 
     private lateinit var parent: ViewGroup
 
@@ -35,39 +36,35 @@ class TextStyleCommandTest {
         parent = LinearLayout(context).apply {
             orientation = LinearLayout.VERTICAL
         }
-        textStyleCommand = TextStyleCommand()
+        textSizeCommand = TextSizeCommand()
 
     }
 
     @Test
-    fun renderingTextShouldApplyTextStyle() {
-
-        val dynamicProperty = DynamicProperty(TextStyleCommand.TEXT_STYLE, "string", "bold")
+    fun renderingTextViewShouldApplyTextSize() {
+        val dynamicProperty = DynamicProperty(TEXT_SIZE, "dimen", "25")
 
         val textView = TextView(parent.context)
-        textStyleCommand.apply(textView, dynamicProperty)
+        textSizeCommand.apply(textView, dynamicProperty)
 
         parent.addView(textView)
 
         textView.afterLayout {
-            assertTrue(textView.typeface.isBold)
+            assertEquals(25f, textView.textSize)
         }
-
     }
 
     @Test
-    fun renderingTextViewWithWrongTextStyleValueShouldThrow() {
+    fun renderingImageViewShouldThrow() {
+        val dynamicProperty = DynamicProperty(TEXT_SIZE, "dimen", "25")
 
-        val dynamicProperty = DynamicProperty(TEXT_STYLE, "string", "wrong")
-
-        val textView = TextView(parent.context)
+        val imageView = ImageView(parent.context)
 
         exceptionRule.expect(IllegalArgumentException::class.java)
-        exceptionRule.expectMessage("The value(${dynamicProperty.value}) " +
-                "for the textStyle property does not exist or is not supported")
+        exceptionRule.expectMessage("The property does not support setting text size" +
+                " for the type ${imageView.javaClass.name}")
 
-        textStyleCommand.apply(textView, dynamicProperty)
-
+        textSizeCommand.apply(imageView, dynamicProperty)
     }
 
 }
