+++ core/src/main/kotlin/br/com/concrete/yosef/api/DynamicViewCreator.kt
@@ -2,7 +2,13 @@ package br.com.concrete.yosef.api
 
 import android.view.ViewGroup
 import br.com.concrete.yosef.OnActionListener
-import br.com.concrete.yosef.api.component.*
+import br.com.concrete.yosef.api.component.Component
+import br.com.concrete.yosef.api.component.ButtonComponent
+import br.com.concrete.yosef.api.component.ElementGroupComponent
+import br.com.concrete.yosef.api.component.RadioButtonComponent
+import br.com.concrete.yosef.api.component.RadioGroupButtonComponent
+import br.com.concrete.yosef.api.component.TextFieldComponent
+import br.com.concrete.yosef.api.component.TextViewComponent
 import br.com.concrete.yosef.api.component.ButtonComponent.Companion.BUTTON_TYPE
 import br.com.concrete.yosef.api.component.ElementGroupComponent.Companion.ELEMENT_GROUP
 import br.com.concrete.yosef.api.component.RadioButtonComponent.Companion.RADIO_BUTTON
