+++ core/src/main/kotlin/br/com/concrete/yosef/api/property/id/IdCommand.kt
