+++ core/src/main/kotlin/br/com/concrete/yosef/api/property/text/TextCommand.kt
