+++ glide/src/main/kotlin/br/com/concrete/yosef/glide/DummyContentProvider.kt
