+++ lottie/src/androidTest/kotlin/br/com/concrete/yosef/lottie/property/AnimatePropertyCommandTest.kt
