--- picasso/src/debug/AndroidManifest.xml
@@ -1,11 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="br.com.concrete.yosef.picasso">
-
-    <application>
-        <provider
-            android:name=".DummyContentProvider"
-            android:authorities="br.com.concrete.yosef.yousefpicasso.provider"
-            android:exported="false" />
-    </application>
-
-</manifest>
