+++ picasso/src/main/kotlin/br/com/concrete/yosef/picasso/DummyContentProvider.kt
